refactor: move all profile files to HR/profile directory
